changed make file + added sdbmc.c file -> will be the main file for execution
